requirements.txt
# Core data
pandas
numpy
scipy

# Machine learning
scikit-learn
xgboost

# Visualisation
matplotlib
seaborn

# Model persistence
joblib

# Jupyter
jupyter
ipykernel

# API
fastapi
uvicorn[standard]
slowapi
jinja2
aiofiles
python-multipart
sqlalchemy

# Authentication
passlib
python-jose